% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_processing_data.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Tidying the raw NOAA dataset}
\usage{
eq_clean_data(df)
}
\arguments{
\item{df}{The raw dataset from NOAA website.}
}
\value{
This function returns the cleaned dataset as in the description.
}
\description{
This function takes raw NOAA data frame and returns a clean data frame.
The clean data frame have the following:
1. A date column created by uniting the year, month, day and converting it to the \code{Date} class
2. \code{LATITUDE} and \code{LONGITUDE} columns converted to \code{numeric} class
}
\examples{
\dontrun{eq_tidy_data <- eq_clean_data(eq_download_data())}

}
