% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_processing_data.R
\name{eq_convert_date}
\alias{eq_convert_date}
\title{Converting BC dates}
\usage{
eq_convert_date(y, m, d)
}
\arguments{
\item{y}{Year}

\item{m}{Month}

\item{d}{Day}
}
\value{
This function returns the properly formatted and combined date of year, month and day
for both BC or AD dates
}
\description{
This function helps properly format BC dates in the dataset.
}
\examples{
\dontrun{eq_convert_date(-2150, 1, 1)}

}
