% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Create labels for earthquake epicenter mapping}
\usage{
eq_create_label(df)
}
\arguments{
\item{df}{The cleaned data set by applying \code{eq_clean_data()} function to the raw data}
}
\description{
This function takes the dataset as an argument and creates an HTML label that can be used as the
annotation text in the leaflet map. This function puts together a character string for each earthquake that will
show the cleaned location, the magnitude (EQ_PRIMARY), and the total number of deaths (TOTAL_DEATHS), with boldface
labels for each ("Location", "Total deaths", and "Magnitude"). If an earthquake is missing values for any of these,
both the label and the value will be skipped for that element of the tag.
}
\examples{
 \dontrun{eq_clean_data(eq_download_data()) \%>\%
                    dplyr::filter(COUNTRY == "Mexico:" & lubridate::year(DATE) >= 2000) \%>\%
                    eq_create_label() \%>\%
                    eq_map(annot_col = "popup_text")}

}
