% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_processing_data.R
\name{eq_download_data}
\alias{eq_download_data}
\title{Download data from NOAA website}
\usage{
eq_download_data()
}
\value{
This function returns the downloaded dataset in a tibble.
}
\description{
This function downloads required dataset from NOAA website.
}
