% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_processing_data.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Cleaning LOCATION_NAME column of the raw NOAA dataset}
\usage{
eq_location_clean(df)
}
\arguments{
\item{df}{The raw dataset from NOAA website.}
}
\value{
This function returns the cleaned dataset as in the description.
}
\description{
This function cleans the LOCATION_NAME column by stripping out the country name (including the colon)
and converts names to title case (as opposed to all caps). This will be needed later for annotating visualizations.
}
\examples{
\dontrun{eq_tidy_data <- eq_location_clean(eq_download_data())}

}
