% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map.R
\name{eq_map}
\alias{eq_map}
\title{Map the earthquake epicenters with pup-up labels}
\usage{
eq_map(df, annot_col = "DATE")
}
\arguments{
\item{df}{The cleaned data set by applying \code{eq_clean_data()} function to the raw data}

\item{annot_col}{The column user wish to use for annotation, default value is "DATE"}
}
\description{
This function takes an argument data containing the filtered data frame with earthquakes to visualize.
The function maps the epicenters (LATITUDE/LONGITUDE) and annotates each point with in pop up window containing
annotation data stored in a column of the data frame. The user is able to choose which column is used for the annotation
in the pop-up with a function argument named annot_col.
}
\examples{
 \dontrun{eq_clean_data(eq_download_data()) \%>\%
                    dplyr::filter(COUNTRY == "Mexico:" & lubridate::year(DATE) >= 2000) \%>\%
                    eq_map(annot_col = "DATE")}

}
