% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_geom_timline.R
\name{geom_timeline}
\alias{geom_timeline}
\title{Plot a timeline of earthquakes within a given time frame}
\usage{
geom_timeline(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = TRUE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Aesthetic mappings created by aes}

\item{data}{Dataset from NOAA website}

\item{stat}{The statistical transformation used}

\item{position}{Position adjustment function}

\item{na.rm}{Remove the NA values from the data frame}

\item{show.legend}{Legend}

\item{inherit.aes}{Default aesthetics}

\item{...}{Other arguments}
}
\value{
The timeline plot.
}
\description{
This function utilizes the \code{GeomTimeline} geom to plot the earthquake events.
In the provided example, the \code{color} aes reflects the number of deaths resulted from the earthquake
with grey ones indicating no data available. The \code{size} aes reflects the magnitude of the eqrthquake.
The X-axis is the date.
}
\examples{
#Pick data points ranging from date \\code{xmin} to \\code{xmax} from the cleaned dataset to plot
           \dontrun{eq_tidy_data <- eq_clean_data(eq_download_data())}
           \dontrun{xmin <- as.Date("1999-01-01")}
           \dontrun{xmax <- as.Date("2000-01-01")}
           \dontrun{sample_data <- filter(eq_tidy_data, DATE >= xmin & DATE <= xmax)}
           \dontrun{ggplot(sample_data,
                   aes(x = DATE, size = EQ_PRIMARY, fill = DEATHS, color = DEATHS)) +
                   geom_timeline() +
                   theme_minimal()}

}
