% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_geom_timeline_label.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{Plot a labeled timeline of earthquakes within a given time frame}
\usage{
geom_timeline_label(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = TRUE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Aesthetic mappings created by aes}

\item{data}{Dataset from NOAA website}

\item{stat}{The statistical transformation used}

\item{position}{Position adjustment function}

\item{na.rm}{Remove the NA values from the data frame}

\item{show.legend}{Legend}

\item{inherit.aes}{Default aesthetics}

\item{...}{Other arguments}
}
\value{
The labeled timeline plot.
}
\description{
This function utilizes the \code{GeomTimelineLable} geom to plot the earthquake events with
labels (location of the earthquake). There is an option to subset to n_max number of earthquakes, where we take the n_max
largest (by magnitude) earthquakes. In the provided example, the \code{color} aes reflects the number of deaths resulted from
the earthquake with grey ones indicating no data available. The \code{size} aes reflects the magnitude of the eqrthquake.
The X-axis is the date. n_max can be customized, the default value is 5.
}
\examples{
#Pick data points ranging from date \\code{xmin} to \\code{xmax} from the cleaned dataset to plot
           \dontrun{eq_tidy_data <- eq_clean_data(eq_download_data())}
           \dontrun{xmin <- as.Date("1999-01-01")}
           \dontrun{xmax <- as.Date("1999-07-01")}
           \dontrun{sample_data <- filter(eq_tidy_data, DATE >= xmin & DATE <= xmax)}
           \dontrun{ggplot(sample_data,
                   aes(x = DATE, label = LOCATION, size = EQ_PRIMARY, fill = DEATHS,
                   color = DEATHS, n_max = 4)) +
                   geom_timeline_label() +
                   theme_minimal()}

}
