% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_geom_timline.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{Plot a timeline of earthquakes}
\format{An object of class \code{GeomTimeline} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.}
\usage{
GeomTimeline
}
\description{
This geom plots a time line of earthquakes with a point for each earthquake.
Optional aesthetics include color, size, and alpha. The x aesthetic is a date when the earthquake occurred.
This geom is the prototype required for \code{geom_timeline()} function to work.
}
\keyword{datasets}
