% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_geom_timeline_label.R
\docType{data}
\name{GeomTimelineLabel}
\alias{GeomTimelineLabel}
\title{Plot a timeline of earthquakes with labels}
\format{An object of class \code{GeomTimelineLable} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.}
\usage{
GeomTimelineLabel
}
\description{
This geom plots a time line of earthquakes with a point for each earthquake.
This geom adds a vertical line to each data point with a text annotation (the location of the earthquake)
attached to each line. There is an option to subset to n_max number of earthquakes, where we take the n_max
largest (by magnitude) earthquakes. Aesthetics are x, which is the date of the earthquake and label which takes
the location column from which annotations will be obtained.
This geom is the prototype required for \code{geom_timeline_label()} function to work.
}
\keyword{datasets}
